Move track pages to apply, move shared stuff into components
