+++ deployment/default.conf
@@ -24,6 +24,22 @@ server {
         root   /usr/share/nginx/html;
     }
 
+    location = /tracks/software-engineering/ {
+        return 301 /apply/software-engineering/;
+    }
+
+    location = /tracks/product-management/ {
+        return 301 /apply/product-management/;
+    }
+
+    location = /tracks/interaction-design/ {
+        return 301 /apply/interaction-design/;
+    }
+
+    location = /tracks/artificial-intelligence/ {
+        return 301 /apply/artificial-intelligence/;
+    }
+
     # proxy the PHP scripts to Apache listening on 127.0.0.1:80
     #
     #location ~ \.php$ {
