+++ src/components/shared/cookie-warning/index.tsx
