+++ src/components/shared/footer/about/_styles.css
