+++ src/components/tracks/offer-responsibilities/_index.tsx
