+++ src/components/tracks/offer-responsibilities/_styles.css
