+++ src/components/tracks/track-image/_styles.css
