+++ src/pages/apply/artificial-intelligence/_styles.css
