+++ src/pages/home/index.tsx
@@ -10,9 +10,9 @@ import Adac from './adac';
 import ToolPartner from './tool-partner'
 import IndustryPartners from './industry-partners';
 import WhatsNew from './whats-new';
-import PickTrack from '../shared/pick-track';
-import Footer from '../shared/footer/index';
-import CookieWarning from '../shared/cookie-warning';
+import PickTrack from '../../components/shared/pick-track';
+import Footer from '../../components/shared/footer/index';
+import CookieWarning from '../../components/shared/cookie-warning';
 
 
 class Home extends React.Component {
